Add benchmark suite, template example and rewritten README
Benchmarks (task 10):
- New dependency-free, header-only benchmark harness
  (benchmark/lib/Benchmark.hpp), mirroring the test harness philosophy.
- Benchmarks one full predict/update cycle for all four filter variants on
  the Robot1 models; emits a plain or Markdown report.
- A Benchmark CI workflow runs the suite and, on pushes to master,
  regenerates the results table in the README via
  scripts/update_readme_benchmark.py.
- Gated behind the KALMAN_BUILD_BENCHMARKS option.

Documentation & adaptability (task 2):
- Rewrite the README: clearer positioning, explicit credit to the original
  mherb/kalman project and what this review changes, quick-start, a 60-second
  example, a guide to adapting the four model pieces, a filter-variant table,
  the benchmark section, and the coding conventions.
- Add examples/Template: a minimal, heavily-commented 1-D constant-velocity
  tracker as a copy-paste starting point.

Also cast Eigen compile-time sizes to int in the SR-EKF augmented matrix to
silence a C++20 enum-arithmetic deprecation warning (surfaced by the
benchmark instantiating that code path).

Builds warning-free and passes all 20 tests under GCC 13 and Clang 18.

+++ include/kalman/SquareRootExtendedKalmanFilter.hpp
@@ -226,7 +226,9 @@ protected:
       const Jacobian<Type, Type> &B, const CovarianceSquareRoot<Type> &R,
       CovarianceSquareRoot<Type> &S_pred) {
     // Compute QR decomposition of (transposed) augmented matrix
-    Matrix<T, State::RowsAtCompileTime + Type::RowsAtCompileTime,
+    Matrix<T,
+           static_cast<int>(State::RowsAtCompileTime) +
+               static_cast<int>(Type::RowsAtCompileTime),
            Type::RowsAtCompileTime>
         tmp;
     tmp.template topRows<State::RowsAtCompileTime>() =
